// packages/shared/src/engine/spirit-field/types.ts
export interface SpiritSeedSkeleton { rank: Quality; quantity: number; forcedElement?: ElementType; regionTags?: string[] }
export interface SpiritSeedRandomOptions { guaranteedRank?: Quality; specifiedElement?: ElementType; regionTags?: string[]; qualityChanceMap?: Record<Quality, number>; rankRange?: { min: Quality; max: Quality } }